requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
wikipedia
googlesearch
youtube_transcript_api
pathlib